pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.1",
    "langchain>=1.2.8",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.9",
    "langchain-text-splitters>=1.1.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.54.0",
]
